+++ src/rsResearch/publisher.cpp
@@ -25,13 +25,13 @@ Publisher::~Publisher(void) {
 /**********************************************************
 	public functions
  **********************************************************/
-int Publisher::send(short id, unsigned int time, float angle) {
+short Publisher::send(short id, unsigned int time, float angle) {
 	char buffer[256];
 	sprintf(buffer, "hard%02d: %u %f", id, time, angle);
 	return zmq_send(_socket, buffer, strlen(buffer), 0);
 }
 
-int Publisher::send(short id, float v, float r, float phi) {
+short Publisher::send(short id, float v, float r, float phi) {
 	char buffer[256];
 	sprintf(buffer, "cpg%02d: %f %f %f", id, v, r, phi);
 	return zmq_send(_socket, buffer, strlen(buffer), 0);
